// PROJECT-OP/src/app/interfaces/service-center.ts
import { Time } from '@angular/common';

export interface ServiveCenterInterface {
  M_Id: number;
  Id: string;
  Source: string;
  Time1: Time;
  Time2: Time;
  Time3: Time;
  Time4: Time;
  Time5: Time;
  Time6: Time;
  Time7: Time;
  Time8: Time;
  Time9: Time;
  Time10: Time;
  Time11: Time;
  Time12: Time;
  Time13: Time;
  Time14: Time;
  Time15: Time;
  Time16: Time;
  TimeRangeId: number;
  Type_ID:number;
}